Done backend with seeder

+++ server/controllers/HomeController.js
@@ -1,18 +1,10 @@
-const { banner, impact } = require('../models');
+const { banner, impact, testimonial, social } = require('../models');
 const createError = require('http-errors');
 const multer = require('multer');
-const path = require('path');
-const serverUrl = 'http://localhost:3000/' // jangan lupa ganti
 const fs = require('fs');
+const storage = require('../helpers/multerStorage');
 
-const storage = multer.diskStorage({
-  destination: function(req, file, cb) {
-    cb(null, 'public/images/')
-  },
-  filename: function(req, file, cb) {
-    cb(null, Date.now() + path.extname(file.originalname))
-  }
-});
+const serverUrl = 'http://localhost:3000/' // jangan lupa ganti
 
 class HomeController {
   static getHomeBanner = async(req, res, next) => {
@@ -25,14 +17,14 @@ class HomeController {
   }
   static editHomeBanner = async(req, res, next) => {
     try {
-      const { title, subtitle } = req.body;
       const upload = multer({ storage: storage }).single('banner_path');
       upload(req, res, async(err) => {
         try {
+          const { title, subtitle } = req.body;
           if (err instanceof multer.MulterError) throw createError(500, 'Internal Server Error');
           else if (err) throw createError(500, 'Internal Server Error');
           if (!req.file) throw createError(400, 'Input Image');
-          const banner_path = serverUrl + req.file.path;
+          const banner_path = serverUrl + req.file.path.replace('public/', '');
           const query = { title, subtitle, banner_path };
           const prevBanner = await banner.findOne({ where: { name: 'Top Banner' } });
           await banner.update(query, { where: { name: 'Top Banner' } });
@@ -59,14 +51,14 @@ class HomeController {
   }
   static editWhoWeAre = async(req, res, next) => {
     try {
-      const { title, subtitle } = req.body;
       const upload = multer({ storage: storage }).single('banner_path');
       upload(req, res, async(err) => {
         try {
+          const { title, subtitle } = req.body;
           if (err instanceof multer.MulterError) throw createError(500, 'Internal Server Error');
           else if (err) throw createError(500, 'Internal Server Error');
           if (!req.file) throw createError(400, 'Input Image');
-          const banner_path = serverUrl + req.file.path;
+          const banner_path = serverUrl + req.file.path.replace('public/', '');
           const query = { title, subtitle, banner_path };
           const prevBanner = await banner.findOne({ where: { name: 'Who We Are' } });
           await banner.update(query, { where: { name: 'Who We Are' } });
@@ -106,10 +98,10 @@ class HomeController {
           else if (err) throw createError(500, 'Internal Server Error');
           let logo_path;
           let main_image_path;
-          if (!req.files['logo_path']) createError(400, 'Input logo image');
           if (!req.files['main_image_path']) createError(400, 'Input main image');
-          if (req.files['logo_path']) logo_path = serverUrl + req.files['logo_path'][0].path;
-          if (req.files['main_image_path']) main_image_path = serverUrl + req.files['main_image_path'][0].path;
+          if (req.files['logo_path']) logo_path = serverUrl + req.files['logo_path'][0].path.replace('public/', '');
+          else logo_path = '';
+          if (req.files['main_image_path']) main_image_path = serverUrl + req.files['main_image_path'][0].path.replace('public/', '');
           const result = await impact.create({ logo_path, main_image_path });
           res.status(201).json(result);
         } catch (err) {
@@ -139,16 +131,16 @@ class HomeController {
           else if (err) throw createError(500, 'Internal Server Error');
           let logo_path;
           let main_image_path;
-          if (!req.files['logo_path']) createError(400, 'Input logo image');
           if (!req.files['main_image_path']) createError(400, 'Input main image');
-          if (req.files['logo_path']) logo_path = serverUrl + req.files['logo_path'][0].path;
-          if (req.files['main_image_path']) main_image_path = serverUrl + req.files['main_image_path'][0].path;
+          if (req.files['logo_path']) logo_path = serverUrl + req.files['logo_path'][0].path.replace('public/', '');
+          else logo_path = '';
+          if (req.files['main_image_path']) main_image_path = serverUrl + req.files['main_image_path'][0].path.replace('public/', '');
           await impact.update({ logo_path, main_image_path }, { where: { id } });
           const regex = new RegExp(serverUrl, "g");
           const prevPathLogo = impactData.logo_path.replace(regex, '');
           const prevPathMain = impactData.main_image_path.replace(regex, '');
-          fs.unlinkSync(prevPathLogo)
-          fs.unlinkSync(prevPathMain)
+          if (impactData.logo_path) fs.unlinkSync(prevPathLogo);
+          fs.unlinkSync(prevPathMain);
           res.status(201).json({ msg: 'Success' });
         } catch (err) {
           if (req.files['main_image_path']) fs.unlinkSync(req.files['main_image_path'][0].path);
@@ -169,8 +161,134 @@ class HomeController {
       const regex = new RegExp(serverUrl, "g");
       const prevPathLogo = impactData.logo_path.replace(regex, '');
       const prevPathMain = impactData.main_image_path.replace(regex, '');
-      fs.unlinkSync(prevPathLogo)
-      fs.unlinkSync(prevPathMain)
+      fs.unlinkSync(prevPathLogo);
+      fs.unlinkSync(prevPathMain);
+      res.status(200).json({ msg: 'Success' });
+    } catch (err) {
+      next(err);
+    }
+  }
+  static getSingleImpact = async(req, res, next) => {
+    try {
+      const { id } = req.params;
+      const result = await impact.findOne({ where: { id } });
+      if (!result) throw createError(404, 'Infograph not found');
+      res.status(200).json(result);
+    } catch (err) {
+      next(err);
+    }
+  }
+  static getAllTestimonial = async(req, res, next) => {
+    try {
+      const result = await testimonial.findAll();
+      res.status(200).json(result);
+    } catch (err) {
+      next(err);
+    }
+  }
+  static createTestimonial = async(req, res, next) => {
+    try {
+      const upload = multer({ storage: storage }).single('photo_path');
+      upload(req, res, async err => {
+        try {
+          if (err instanceof multer.MulterError) throw createError(500, 'Internal Server Error');
+          else if (err) throw createError(500, 'Internal Server Error');
+          const { title, message, job_description } = req.body;
+          if (!title || !message || !job_description) throw createError(400, 'Input all forms');
+          const photo_path = serverUrl + req.file.path.replace('public/', '');
+          const result = await testimonial.create({
+            title,
+            message,
+            job_description,
+            photo_path
+          });
+          res.status(201).json(result);
+        } catch (err) {
+          if (req.file) fs.unlinkSync(req.file.path);
+          next(err);
+        }
+      });
+    } catch (err) {
+      next(err);
+    }
+  }
+  static getSingleTestimonial = async(req, res, next) => {
+    try {
+      const { id } = req.params;
+      const result = await testimonial.findOne({ where: { id } });
+      if (!result) throw createError(404, 'Testimonial not found');
+      res.status(200).json(result);
+    } catch (err) {
+      next(err);
+    }
+  }
+  static updateTestimonial = async(req, res, next) => {
+    try {
+      const { id } = req.params;
+      const testimonialData = await testimonial.findOne({ where: { id } });
+      if (!testimonialData) throw createError(404, 'Testimonial not found');
+      const upload = multer({ storage: storage }).single('photo_path');
+      upload(req, res, async err => {
+        try {
+          if (err instanceof multer.MulterError) throw createError(500, 'Internal Server Error');
+          else if (err) throw createError(500, 'Internal Server Error');
+          const { title, message, job_description } = req.body;
+          if (!title || !message || !job_description) throw createError(400, 'Input all forms');
+          if (!id) throw createError(404, 'Input Id in params')
+          const photo_path = serverUrl + req.file.path.replace('public/', '');
+          await testimonial.update({
+            title,
+            message,
+            job_description,
+            photo_path
+          }, {
+            where: {
+              id
+            }
+          });
+          const regex = new RegExp(serverUrl, "g");
+          const prevPath = testimonialData.photo_path.replace(regex, '');
+          fs.unlinkSync(prevPath);
+          res.status(200).json({ msg: 'Success' });
+        } catch (err) {
+          if (req.file) fs.unlinkSync(req.file.path);
+          next(err);
+        }
+      });
+    } catch (err) {
+      next(err);
+    }
+  }
+  static deleteTestimonial = async(req, res, next) => {
+    try {
+      const { id } = req.params;
+      const testimonialData = await testimonial.findOne({ where: { id } });
+      if (!testimonialData) throw createError(404, 'Testimonial not found');
+      await testimonial.destroy({ where: { id } });
+      const regex = new RegExp(serverUrl, "g");
+      const prevPath = testimonialData.photo_path.replace(regex, '');
+      fs.unlinkSync(prevPath);
+      res.status(200).json({ msg: 'Success' });
+    } catch (err) {
+      next(err);
+    }
+  }
+  static getAllSocial = async(req, res, next) => {
+    try {
+      const result = await social.findAll();
+      res.status(200).json(result);
+    } catch (err) {
+      next(err);
+    }
+  }
+  static editSocial = async(req, res, next) => {
+    try {
+      const { id } = req.params;
+      const socialData = await social.findOne({ where: { id } });
+      if (!socialData) throw createError(404, 'Social Media not found');
+      const { link } = req.body;
+      if (!link) throw createError(400, 'Input a link');
+      await social.update({ link }, { where: { id } });
       res.status(200).json({ msg: 'Success' });
     } catch (err) {
       next(err);
